requirements.txt
django
gunicorn
django-heroku
django-environ
requests
whitenoise
djangorestframework
beautifulsoup4